# mod-api-openapi.json
{
  "components": {
    "schemas": {
      "Connection": {
        "description": "Represents an audio/MIDI connection between ports.\n\nA connection links an input port to an output port, enabling signal flow\nthrough the pedalboard.",
        "properties": {
          "id": {
            "default": 0,
            "description": "Unique identifier for this connection",
            "title": "Id",
            "type": "integer"
          },
          "input_port_id": {
            "description": "ID of the input port (receiving signal)",
            "title": "Input Port Id",
            "type": "string"
          },
          "output_port_id": {
            "description": "ID of the output port (sending signal)",
            "title": "Output Port Id",
            "type": "string"
          }
        },
        "required": [
          "input_port_id",
          "output_port_id"
        ],
        "title": "Connection",
        "type": "object"
      },
...
      "CreatePedalboardRequest": {
        "description": "Request body for creating a pedalboard.",
        "properties": {
          "duplicate_current": {
            "default": false,
            "description": "If True, duplicate the current pedalboard's effects and connections",
            "title": "Duplicate Current",
            "type": "boolean"
          },
          "name": {
            "description": "Human-readable name for the new pedalboard",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CreatePedalboardRequest",
        "type": "object"
      },
...
      "EffectInstance": {
        "description": "Represents an instantiated effect plugin on a pedalboard.\n\nInherits from EffectInfo for ports and other metadata, but overrides parameters\nto store current values instead of definitions.",
        "properties": {
          "id": {
            "default": 0,
            "description": "Unique identifier for this effect instance",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Human-readable plugin name",
            "title": "Name"
          },
          "parameters": {
            "additionalProperties": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/NumberParameter"
                },
                {
                  "$ref": "#/components/schemas/FilenameParameter"
                }
              ]
            },
            "description": "Dictionary mapping parameter names to Parameter objects with values",
            "title": "Parameters",
            "type": "object"
          },
          "ports": {
            "description": "List of input/output ports on the plugin",
            "items": {
              "$ref": "#/components/schemas/Port"
            },
            "title": "Ports",
            "type": "array"
          },
          "uri": {
            "description": "Full URI of the LV2 plugin",
            "title": "Uri",
            "type": "string"
          }
        },
        "required": [
          "uri"
        ],
        "title": "EffectInstance",
        "type": "object"
      },
      "FilenameParameter": {
        "description": "Filename parameter instance with a current value.\n\nExtends FilenameParameterType by adding a current value field.",
        "properties": {
          "default": {
            "description": "Default file path",
            "title": "Default",
            "type": "string"
          },
          "name": {
            "description": "The parameter name/symbol",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "const": "filename",
            "default": "filename",
            "title": "Type",
            "type": "string"
          },
          "value": {
            "description": "Current file path value",
            "title": "Value",
            "type": "string"
          }
        },
        "required": [
          "name",
          "default",
          "value"
        ],
        "title": "FilenameParameter",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "NumberParameter": {
        "description": "Numeric parameter instance with a current value.\n\nExtends NumberParameterType by adding a current value field.",
        "properties": {
          "default": {
            "description": "Default value for the parameter",
            "title": "Default",
            "type": "number"
          },
          "max": {
            "description": "Maximum allowed value",
            "title": "Max",
            "type": "number"
          },
          "min": {
            "description": "Minimum allowed value",
            "title": "Min",
            "type": "number"
          },
          "name": {
            "description": "The parameter name/symbol",
            "title": "Name",
            "type": "string"
          },
          "type": {
            "const": "number",
            "default": "number",
            "title": "Type",
            "type": "string"
          },
          "value": {
            "description": "Current parameter value",
            "title": "Value",
            "type": "number"
          }
        },
        "required": [
          "name",
          "min",
          "max",
          "default",
          "value"
        ],
        "title": "NumberParameter",
        "type": "object"
      },
...
      "Pedalboard": {
        "description": "Represents a complete pedalboard configuration.\n\nA pedalboard contains effect instances and their connections.\nParameter values live inside each effect's parameters field.",
        "properties": {
          "connections": {
            "additionalProperties": {
              "$ref": "#/components/schemas/Connection"
            },
            "description": "Dict of connections keyed by connection ID",
            "title": "Connections",
            "type": "object"
          },
          "effects": {
            "additionalProperties": {
              "$ref": "#/components/schemas/EffectInstance"
            },
            "description": "Dict of effect instances keyed by instance ID",
            "title": "Effects",
            "type": "object"
          },
          "file": {
            "description": "JSON filename for persistence (derived from name)",
            "title": "File",
            "type": "string"
          },
          "id": {
            "description": "The unique pedalboard identifier",
            "title": "Id",
            "type": "integer"
          },
          "name": {
            "description": "Human-readable pedalboard name",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "file"
        ],
        "title": "Pedalboard",
        "type": "object"
      },
      "Port": {
        "description": "Port representation for audio/MIDI routing.\n\nRepresents an audio or MIDI port that can be used for signal routing\nin a pedalboard. Ports can be system ports (from JACK) or effect ports.",
        "properties": {
          "effect_instance_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The effect instance ID if owner_type is 'effect', None otherwise",
            "title": "Effect Instance Id"
          },
          "name": {
            "description": "Human-readable port name",
            "title": "Name",
            "type": "string"
          },
          "owner_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Either 'system' or 'effect'",
            "title": "Owner Type"
          },
          "type": {
            "description": "The type of port (input/output/midi_in/midi_out)",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type"
        ],
        "title": "Port",
        "type": "object"
      },
      "RenamePedalboardRequest": {
        "description": "Request body for renaming a pedalboard.",
        "properties": {
          "name": {
            "description": "New human-readable name for the pedalboard",
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "RenamePedalboardRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "REST API for managing pedalboards, effects, and connections on MOD Audio platform",
    "title": "Pedalboard API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/pedalboards": {
      "get": {
        "description": "List all available pedalboards with metadata.\n\nReturns a dict of all pedalboards keyed by ID.",
        "operationId": "list_pedalboards_api_v1_pedalboards_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "$ref": "#/components/schemas/Pedalboard"
                  },
                  "title": "Response List Pedalboards Api V1 Pedalboards Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Pedalboards",
        "tags": [
          "Pedalboards"
        ]
      },
      "post": {
        "description": "Create a new pedalboard.\n\nCan create empty or as a duplicate of the current board.",
        "operationId": "create_pedalboard_api_v1_pedalboards_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePedalboardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedalboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Pedalboard",
        "tags": [
          "Pedalboards"
        ]
      }
    },
...
    "/api/v1/pedalboards/{pedalboard_id}": {
      "delete": {
        "description": "Delete a pedalboard.\n\nIf deleting the current pedalboard, no pedalboard will be current\nuntil a new one is selected.\n\nRaises:\n    HTTPException: 404 if pedalboard ID doesn't exist.",
        "operationId": "delete_pedalboard_api_v1_pedalboards__pedalboard_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "pedalboard_id",
            "required": true,
            "schema": {
              "title": "Pedalboard Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Pedalboard",
        "tags": [
          "Pedalboards"
        ]
      },
      "get": {
        "description": "Get full details of a specific pedalboard.\n\nRaises:\n    HTTPException: 404 if pedalboard ID doesn't exist.",
        "operationId": "get_pedalboard_api_v1_pedalboards__pedalboard_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "pedalboard_id",
            "required": true,
            "schema": {
              "title": "Pedalboard Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedalboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pedalboard",
        "tags": [
          "Pedalboards"
        ]
      }
    },
...
    "/api/v1/pedalboards/{pedalboard_id}/rename": {
      "put": {
        "description": "Rename a pedalboard.\n\nRaises:\n    HTTPException: 404 if pedalboard ID doesn't exist.\n    HTTPException: 400 if name is missing or invalid.",
        "operationId": "rename_pedalboard_api_v1_pedalboards__pedalboard_id__rename_put",
        "parameters": [
          {
            "in": "path",
            "name": "pedalboard_id",
            "required": true,
            "schema": {
              "title": "Pedalboard Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenamePedalboardRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pedalboard"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Rename Pedalboard",
        "tags": [
          "Pedalboards"
        ]
      }
    },
    "/api/v1/pedalboards/{pedalboard_id}/select": {
      "put": {
        "description": "Select a pedalboard to make it the current one.\n\nRaises:\n    HTTPException: 404 if pedalboard ID doesn't exist.",
        "operationId": "select_pedalboard_api_v1_pedalboards__pedalboard_id__select_put",
        "parameters": [
          {
            "in": "path",
            "name": "pedalboard_id",
            "required": true,
            "schema": {
              "title": "Pedalboard Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Select Pedalboard",
        "tags": [
          "Pedalboards"
        ]
      }
    },
...
  }
}